job_tracker/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
openpyxl==3.1.2 